Aggiunta e partecipazione lavaggi core

+++ codice/washineCore/src/main/java/washine/washineCore/WashineCoreWashing.java
@@ -17,12 +17,18 @@ import washine_db.washings.WashineWashingDb;
 public class WashineCoreWashing implements WashineCoreWashingIf {
 
   @Override
-  public boolean createWashing(String washingId, String laundryPersonId)
-      throws WashineCoreException {
+  public WashineWashing createWashing(String laundryPersonId) throws WashineCoreException {
     WashineWashingDb washingDb = new WashineWashingDb();
+    String washingId = generateUniqueCode();
 
     try {
-      return washingDb.createWashing(washingId, laundryPersonId);
+      if (washingDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing already exists");
+      }
+      washingDb.createWashing(washingId, laundryPersonId);
+      WashineWashingOptions washingOptions = new WashineWashingOptions();
+      return new WashineWashing(washingId, null, washingOptions);
+
     } catch (WashineDataException e) {
       throw new WashineCoreException("WashineCoreException");
     }
@@ -33,6 +39,9 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
       throws WashineCoreException {
     WashineWashingDb washingDb = new WashineWashingDb();
     try {
+      if (!washingDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing does not exists");
+      }
       Result<WashingoptionsRecord> washingOptionsDb = washingDb.getWashingOptions(washingId);
       WashineWashingOptions washingOptions =
           new WashineWashingOptions(
@@ -67,18 +76,14 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
       if (load > washingOptions.getMaxLoadParticipant()) {
         throw new WashineCoreException("Exceded maximum load");
       }
-
-      double totalWashingLoad = washingOptions.getInitialLoad();
-      List<String> participantIds = washingDb.getParticipantIds(washingId);
-      for (String s : participantIds) {
-        totalWashingLoad += washingDb.getParticipationWeight(washingId, s);
-      }
+      WashineWashing washing = new WashineWashing(washingId, null, washingOptions);
+      double totalWashingLoad = calculateWashingTotalLoad(washing);
 
       if (totalWashingLoad >= washingOptions.getMaxLoad()) {
         throw new WashineCoreException("Max load Reached");
       }
       washingDb.participateToWashing(washingId, participantId, load);
-      return new WashineWashing(washingId, null, washingOptions);
+      return washing;
     } catch (WashineDataException e) {
       throw new WashineCoreException("WashineCoreException");
     }
@@ -108,10 +113,12 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
   @Override
   public double getParticipationWeight(String washingId, String userId)
       throws WashineCoreException {
-    WashineWashingDb washineDb = new WashineWashingDb();
+    WashineWashingDb washingDb = new WashineWashingDb();
     try {
-
-      return washineDb.getParticipationWeight(washingId, userId);
+      if (!washingDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing does not exists");
+      }
+      return washingDb.getParticipationWeight(washingId, userId);
     } catch (WashineDataException e) {
 
       throw new WashineCoreException("WashineCoreException");
@@ -120,11 +127,13 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
 
   @Override
   public boolean deleteWashing(String washingId) throws WashineCoreException {
-    WashineWashingDb washineDb = new WashineWashingDb();
+    WashineWashingDb washingDb = new WashineWashingDb();
     try {
-
-      if (washineDb.getWashingOptions(washingId) != null) {
-        return washineDb.deleteWashing(washingId);
+      if (!washingDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing does not exists");
+      }
+      if (washingDb.getWashingOptions(washingId) != null) {
+        return washingDb.deleteWashing(washingId);
       } else {
         return false;
       }
@@ -136,64 +145,74 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
   @Override
   public Result<WashingoptionsRecord> getWashingOptions(String washingId)
       throws WashineCoreException {
-    WashineWashingDb washineDb = new WashineWashingDb();
+    WashineWashingDb washingDb = new WashineWashingDb();
     try {
-      return washineDb.getWashingOptions(washingId);
+      if (!washingDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing does not exists");
+      }
+      return washingDb.getWashingOptions(washingId);
     } catch (WashineDataException e) {
       throw new WashineCoreException("WashineCoreException");
     }
   }
 
   @Override
-  public boolean updateWashingOptions(
-      String washingId,
-      int visibilityTime,
-      int dateTime,
-      int durationMinutes,
-      double initialLoad,
-      double maxLoad,
-      String temperature,
-      String spinSpeed,
-      String fabricType,
-      String color,
-      String detergentTypes,
-      String refundTypes,
-      boolean underwear,
-      String pickupAddress,
-      String deliveryAddress,
-      String pickupAvailability,
-      String deliveryAvailability,
-      String drying,
-      boolean ironing,
-      double participantMaxLoad,
-      int washingAccessOpenDate,
-      int washingAccessCloseDate)
-      throws WashineCoreException {
-    WashineWashingDb washineDb = new WashineWashingDb();
+  public boolean updateWashingOptions(WashineWashing washing) throws WashineCoreException {
+    WashineWashingDb washingDb = new WashineWashingDb();
     try {
-      return washineDb.updateWashingOptions(
-          washingId,
-          visibilityTime,
-          dateTime,
-          durationMinutes,
-          initialLoad,
-          maxLoad,
-          temperature,
-          spinSpeed,
-          fabricType,
-          color,
-          detergentTypes,
-          refundTypes,
-          underwear,
-          pickupAddress,
-          deliveryAddress,
-          pickupAvailability,
-          deliveryAvailability,
-          drying,
-          ironing,
-          participantMaxLoad,
-          washingAccessOpenDate,
-          washingAccessCloseDate);
+      if (!washingDb.existsWashing(washing.getId())) {
+        throw new WashineCoreException("Washing does not exists");
+      }
+      if (washing.getWashingOptions().getDatetime() <= 0) {
+        throw new WashineCoreException("Invalid date and time");
+      }
+      if (washing.getWashingOptions().getDurationMinutes() <= 0) {
+        throw new WashineCoreException("Invalid duration minutes");
+      }
+      if (washing.getWashingOptions().getInitialLoad() < 0) {
+        throw new WashineCoreException("Invalid initial load");
+      }
+      if (washing.getWashingOptions().getMaxLoad() < 0) {
+        throw new WashineCoreException("Invalid max load");
+      }
+      if (washing.getWashingOptions().getMaxLoadParticipant() < 0) {
+        throw new WashineCoreException("Invalid participant max load");
+      }
+      if (washing.getWashingOptions().getVisibilityTime() <= 0) {
+        throw new WashineCoreException("Invalid visibility time");
+      }
+      if (washing.getWashingOptions().getWashingAccessCloseDate() <= 0) {
+        throw new WashineCoreException("Invalid access date");
+      }
+      if (washing.getWashingOptions().getWashingAccessCloseDate() <= 0) {
+        throw new WashineCoreException("Invalid closing date");
+      }
+      if (calculateWashingTotalLoad(washing) > washing.getWashingOptions().getMaxLoad()) {
+        throw new WashineCoreException("New max load exceeded");
+      }
+      return washingDb.updateWashingOptions(
+          washing.getId(),
+          washing.getWashingOptions().getVisibilityTime(),
+          washing.getWashingOptions().getDatetime(),
+          washing.getWashingOptions().getDurationMinutes(),
+          washing.getWashingOptions().getInitialLoad(),
+          washing.getWashingOptions().getMaxLoad(),
+          washing.getWashingOptions().getTemperature(),
+          washing.getWashingOptions().getSpinSpeed(),
+          washing.getWashingOptions().getFabricType(),
+          washing.getWashingOptions().getColor(),
+          washing.getWashingOptions().getDetergentType(),
+          washing.getWashingOptions().getRefundType(),
+          washing.getWashingOptions().isUnderwear(),
+          washing.getWashingOptions().getPickupAddress(),
+          washing.getWashingOptions().getDeliveryAddress(),
+          washing.getWashingOptions().getPickupAvailability(),
+          washing.getWashingOptions().getDeliveryAvailability(),
+          washing.getWashingOptions().getDryingType(),
+          washing.getWashingOptions().isIroning(),
+          washing.getWashingOptions().getMaxLoadParticipant(),
+          washing.getWashingOptions().getWashingAccessOpenDate(),
+          washing.getWashingOptions().getWashingAccessCloseDate());
     } catch (WashineDataException e) {
       throw new WashineCoreException("WashineCoreException");
     }
@@ -203,13 +222,20 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
   public boolean deleteParticipation(String washingId, String userId) throws WashineCoreException {
     WashineWashingDb washineDb = new WashineWashingDb();
     try {
+      if (!washineDb.existsWashing(washingId)) {
+        throw new WashineCoreException("Washing does not exists");
+      }
       return washineDb.deleteParticipation(washingId, userId);
+
     } catch (WashineDataException e) {
 
       throw new WashineCoreException("WashineCoreException");
     }
   }
 
+  /**
+   * @return a new generated and unique id code
+   */
   public String generateUniqueCode() {
     String SALTCHARS = "ABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890";
     StringBuilder salt = new StringBuilder();
@@ -219,5 +245,27 @@ public class WashineCoreWashing implements WashineCoreWashingIf {
       salt.append(SALTCHARS.charAt(index));
     }
     return salt.toString();
+  }
+
+  /**
+   * Calculates the total load of a washing through the id
+   *
+   * @param washing the washing object (cointaing the supposed id)
+   * @return the total load
+   * @throws WashineCoreException
+   */
+  public double calculateWashingTotalLoad(WashineWashing washing) throws WashineCoreException {
+    WashineWashingDb washingDb = new WashineWashingDb();
+    try {
+      double totalWashingLoad = washing.getWashingOptions().getInitialLoad();
+      List<String> participantIds = washingDb.getParticipantIds(washing.getId());
+      for (String s : participantIds) {
+        totalWashingLoad += washingDb.getParticipationWeight(washing.getId(), s);
+        return totalWashingLoad;
+      }
+    } catch (WashineDataException e) {
+      throw new WashineCoreException("WashineCoreException");
+    }
+    return 0;
   }
 }
